Durschnuttsverbrauch double anstelle int

+++ src/de/merit/azubi/Durchschnittsverbrauch.java
@@ -5,7 +5,7 @@ public class Durchschnittsverbrauch {
     public static void main(String[] args) {
 
         int gefKm = 430;
-        int verbrauchL = 38;
+        double verbrauchL = 38;
 
         double erg = verbrauchL / gefKm * 100;
